(#471) Pass params to snapshot objects and set snapshot microns based on these. Also move OAVConfigParams
