+++ src/dodal/devices/oav/oav_detector.py
@@ -1,6 +1,4 @@
-import xml.etree.cElementTree as et
 from functools import partial
-from typing import Tuple
 
 from ophyd import ADComponent as ADC
 from ophyd import (
@@ -18,17 +16,7 @@ from ophyd import (
 )
 
 from dodal.devices.oav.grid_overlay import SnapshotWithGrid
-from dodal.devices.oav.oav_errors import (
-    OAVError_BeamPositionNotFound,
-    OAVError_ZoomLevelNotFound,
-)
-from dodal.log import LOGGER
-
-# GDA corrently assumes this aspect ratio for the OAV window size.
-# For some beamline this doesn't affect anything as the actual OAV aspect ratio
-# matches. Others need to take it into account to rescale the values stored in
-# the configuration files.
-DEFAULT_OAV_WINDOW = (1024, 768)
+from dodal.devices.oav.oav_parameters import OAVConfigParams
 
 
 class ZoomController(Device):
@@ -82,111 +70,6 @@ class ZoomController(Device):
         return return_status
 
 
-class OAVConfigParams:
-    """
-    The OAV parameters which may update depending on settings such as the zoom level.
-    """
-
-    def __init__(
-        self,
-        zoom_params_file,
-        display_config,
-    ):
-        self.zoom_params_file: str = zoom_params_file
-        self.display_config: str = display_config
-
-    def update_on_zoom(self, value, xsize, ysize, *args, **kwargs):
-        xsize, ysize = int(xsize), int(ysize)
-        if isinstance(value, str) and value.endswith("x"):
-            value = value.strip("x")
-        zoom = float(value)
-        self.load_microns_per_pixel(zoom, xsize, ysize)
-        self.beam_centre_i, self.beam_centre_j = self.get_beam_position_from_zoom(
-            zoom, xsize, ysize
-        )
-
-    def load_microns_per_pixel(self, zoom: float, xsize: int, ysize: int) -> None:
-        """
-        Loads the microns per x pixel and y pixel for a given zoom level. These are
-        currently generated by GDA, though hyperion could generate them in future.
-        """
-        tree = et.parse(self.zoom_params_file)
-        self.micronsPerXPixel = self.micronsPerYPixel = None
-        root = tree.getroot()
-        levels = root.findall(".//zoomLevel")
-        for node in levels:
-            if float(node.find("level").text) == zoom:
-                self.micronsPerXPixel = (
-                    float(node.find("micronsPerXPixel").text)
-                    * DEFAULT_OAV_WINDOW[0]
-                    / xsize
-                )
-                self.micronsPerYPixel = (
-                    float(node.find("micronsPerYPixel").text)
-                    * DEFAULT_OAV_WINDOW[1]
-                    / ysize
-                )
-        if self.micronsPerXPixel is None or self.micronsPerYPixel is None:
-            raise OAVError_ZoomLevelNotFound(
-                f"""
-                Could not find the micronsPer[X,Y]Pixel parameters in
-                {self.zoom_params_file} for zoom level {zoom}.
-                """
-            )
-
-    def get_beam_position_from_zoom(
-        self, zoom: float, xsize: int, ysize: int
-    ) -> Tuple[int, int]:
-        """
-        Extracts the beam location in pixels `xCentre` `yCentre`, for a requested zoom \
-        level. The beam location is manually inputted by the beamline operator on GDA \
-        by clicking where on screen a scintillator lights up, and stored in the \
-        display.configuration file.
-        """
-        crosshair_x_line = None
-        crosshair_y_line = None
-        with open(self.display_config, "r") as f:
-            file_lines = f.readlines()
-            for i in range(len(file_lines)):
-                if file_lines[i].startswith("zoomLevel = " + str(zoom)):
-                    crosshair_x_line = file_lines[i + 1]
-                    crosshair_y_line = file_lines[i + 2]
-                    break
-
-        if crosshair_x_line is None or crosshair_y_line is None:
-            raise OAVError_BeamPositionNotFound(
-                f"Could not extract beam position at zoom level {zoom}"
-            )
-
-        beam_centre_i = (
-            int(crosshair_x_line.split(" = ")[1]) * xsize / DEFAULT_OAV_WINDOW[0]
-        )
-        beam_centre_j = (
-            int(crosshair_y_line.split(" = ")[1]) * ysize / DEFAULT_OAV_WINDOW[1]
-        )
-        LOGGER.info(f"Beam centre: {beam_centre_i, beam_centre_j}")
-        return int(beam_centre_i), int(beam_centre_j)
-
-    def calculate_beam_distance(
-        self, horizontal_pixels: int, vertical_pixels: int
-    ) -> Tuple[int, int]:
-        """
-        Calculates the distance between the beam centre and the given (horizontal, vertical).
-
-        Args:
-            horizontal_pixels (int): The x (camera coordinates) value in pixels.
-            vertical_pixels (int): The y (camera coordinates) value in pixels.
-        Returns:
-            The distance between the beam centre and the (horizontal, vertical) point in pixels as a tuple
-            (horizontal_distance, vertical_distance).
-        """
-
-        return (
-            self.beam_centre_i - horizontal_pixels,
-            self.beam_centre_j - vertical_pixels,
-        )
-
-
 class OAV(AreaDetector):
     cam = ADC(CamBase, "-DI-OAV-01:CAM:")
     roi = ADC(ROIPlugin, "-DI-OAV-01:ROI:")
@@ -195,12 +78,12 @@ class OAV(AreaDetector):
     tiff = ADC(OverlayPlugin, "-DI-OAV-01:TIFF:")
     hdf5 = ADC(HDF5Plugin, "-DI-OAV-01:HDF5:")
     grid_snapshot = Component(SnapshotWithGrid, "-DI-OAV-01:MJPG:")
-
     zoom_controller = Component(ZoomController, "-EA-OAV-01:FZOOM:")
 
     def __init__(self, *args, params: OAVConfigParams, **kwargs):
         super().__init__(*args, **kwargs)
         self.parameters = params
+        self.grid_snapshot.oav_params = params
         self.subscription_id = None
         self._snapshot_trigger_subscription_id = None
 
@@ -215,20 +98,4 @@ class OAV(AreaDetector):
             self.zoom_controller.level.unsubscribe(self.subscription_id)
         self.subscription_id = self.zoom_controller.level.subscribe(cb)
 
-        self.add_snapshot_trigger_hook()
-
         return connected
-
-    def add_snapshot_trigger_hook(self):
-        def apply_current_microns_per_pixel_to_snapshot(*args, **kwargs):
-            """Persist the current value of the mpp to the snapshot so that it is retained if zoom changes"""
-            microns_per_x_pixel = self.parameters.micronsPerXPixel
-            microns_per_y_pixel = self.parameters.micronsPerYPixel
-            self.grid_snapshot.microns_per_pixel_x.put(microns_per_x_pixel)
-            self.grid_snapshot.microns_per_pixel_y.put(microns_per_y_pixel)
-
-        self._snapshot_trigger_subscription_id = (
-            self.grid_snapshot.last_saved_path.subscribe(
-                apply_current_microns_per_pixel_to_snapshot
-            )
-        )
