+++ src/dodal/devices/oav/oav_parameters.py
@@ -1,6 +1,19 @@
 import json
+import xml.etree.cElementTree as et
 from collections import ChainMap
-from typing import Any
+from typing import Any, Tuple
+
+from dodal.devices.oav.oav_errors import (
+    OAVError_BeamPositionNotFound,
+    OAVError_ZoomLevelNotFound,
+)
+from dodal.log import LOGGER
+
+# GDA currently assumes this aspect ratio for the OAV window size.
+# For some beamline this doesn't affect anything as the actual OAV aspect ratio
+# matches. Others need to take it into account to rescale the values stored in
+# the configuration files.
+DEFAULT_OAV_WINDOW = (1024, 768)
 
 OAV_CONFIG_JSON = (
     "/dls_sw/i03/software/daq_configuration/json/OAVCentring_hyperion.json"
@@ -86,3 +99,108 @@ class OAVParameters:
         Get the maximum tip distance in pixels.
         """
         return self.max_tip_distance / micronsPerPixel
+
+
+class OAVConfigParams:
+    """
+    The OAV parameters which may update depending on settings such as the zoom level.
+    """
+
+    def __init__(
+        self,
+        zoom_params_file,
+        display_config,
+    ):
+        self.zoom_params_file: str = zoom_params_file
+        self.display_config: str = display_config
+
+    def update_on_zoom(self, value, xsize, ysize, *args, **kwargs):
+        xsize, ysize = int(xsize), int(ysize)
+        if isinstance(value, str) and value.endswith("x"):
+            value = value.strip("x")
+        zoom = float(value)
+        self.load_microns_per_pixel(zoom, xsize, ysize)
+        self.beam_centre_i, self.beam_centre_j = self.get_beam_position_from_zoom(
+            zoom, xsize, ysize
+        )
+
+    def load_microns_per_pixel(self, zoom: float, xsize: int, ysize: int) -> None:
+        """
+        Loads the microns per x pixel and y pixel for a given zoom level. These are
+        currently generated by GDA, though hyperion could generate them in future.
+        """
+        tree = et.parse(self.zoom_params_file)
+        self.micronsPerXPixel = self.micronsPerYPixel = None
+        root = tree.getroot()
+        levels = root.findall(".//zoomLevel")
+        for node in levels:
+            if float(node.find("level").text) == zoom:
+                self.micronsPerXPixel = (
+                    float(node.find("micronsPerXPixel").text)
+                    * DEFAULT_OAV_WINDOW[0]
+                    / xsize
+                )
+                self.micronsPerYPixel = (
+                    float(node.find("micronsPerYPixel").text)
+                    * DEFAULT_OAV_WINDOW[1]
+                    / ysize
+                )
+        if self.micronsPerXPixel is None or self.micronsPerYPixel is None:
+            raise OAVError_ZoomLevelNotFound(
+                f"""
+                Could not find the micronsPer[X,Y]Pixel parameters in
+                {self.zoom_params_file} for zoom level {zoom}.
+                """
+            )
+
+    def get_beam_position_from_zoom(
+        self, zoom: float, xsize: int, ysize: int
+    ) -> Tuple[int, int]:
+        """
+        Extracts the beam location in pixels `xCentre` `yCentre`, for a requested zoom \
+        level. The beam location is manually inputted by the beamline operator on GDA \
+        by clicking where on screen a scintillator lights up, and stored in the \
+        display.configuration file.
+        """
+        crosshair_x_line = None
+        crosshair_y_line = None
+        with open(self.display_config, "r") as f:
+            file_lines = f.readlines()
+            for i in range(len(file_lines)):
+                if file_lines[i].startswith("zoomLevel = " + str(zoom)):
+                    crosshair_x_line = file_lines[i + 1]
+                    crosshair_y_line = file_lines[i + 2]
+                    break
+
+        if crosshair_x_line is None or crosshair_y_line is None:
+            raise OAVError_BeamPositionNotFound(
+                f"Could not extract beam position at zoom level {zoom}"
+            )
+
+        beam_centre_i = (
+            int(crosshair_x_line.split(" = ")[1]) * xsize / DEFAULT_OAV_WINDOW[0]
+        )
+        beam_centre_j = (
+            int(crosshair_y_line.split(" = ")[1]) * ysize / DEFAULT_OAV_WINDOW[1]
+        )
+        LOGGER.info(f"Beam centre: {beam_centre_i, beam_centre_j}")
+        return int(beam_centre_i), int(beam_centre_j)
+
+    def calculate_beam_distance(
+        self, horizontal_pixels: int, vertical_pixels: int
+    ) -> Tuple[int, int]:
+        """
+        Calculates the distance between the beam centre and the given (horizontal, vertical).
+
+        Args:
+            horizontal_pixels (int): The x (camera coordinates) value in pixels.
+            vertical_pixels (int): The y (camera coordinates) value in pixels.
+        Returns:
+            The distance between the beam centre and the (horizontal, vertical) point in pixels as a tuple
+            (horizontal_distance, vertical_distance).
+        """
+
+        return (
+            self.beam_centre_i - horizontal_pixels,
+            self.beam_centre_j - vertical_pixels,
+        )
